+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.0.0 Under development
 
+- New: Module's class has been renamed from `dektrium\rbac\Module` to `dektrium\rbac\RbacWebModule` (dmeroff)
 - Enh #9: Added ability to manage rules (dmeroff)
 - Enh: Updated grid views (dmeroff)
 - Enh: Added check if rule has name property set (dmeroff)
@@ -17,8 +18,8 @@
 
 ## 0.2.0 December 5, 2015
 
-- Some fixes and improvements
+Some fixes and improvements
 
 ## 0.1.0 June 14, 2015
 
-- Initial release
+Initial release